Fix scinote edit in attachments [SCI-10834]
