+++ app/javascript/vue/shared/content/attachments/attachment_actions.vue
@@ -2,7 +2,6 @@
   <div class="flex items-center gap-1.5 justify-end w-[184px]">
     <OpenMenu
       :attachment="attachment"
-      :multipleOpenOptions="multipleOpenOptions"
       @open="$emit('attachment:toggle_menu', $event)"
       @close="$emit('attachment:toggle_menu', $event)"
       @menu-dropdown-toggle="$emit('attachment:toggle_menu', $event)"
@@ -48,41 +47,6 @@ export default {
   components: {
     OpenMenu,
     ContextMenu
-  },
-  mounted() {
-    this.fetchLocalAppInfo();
-  },
-  computed: {
-    multipleOpenOptions() {
-      const options = [];
-      if (this.attachment.attributes.wopi && this.attachment.attributes.urls.edit_asset) {
-        options.push({
-          text: this.attachment.attributes.wopi_context.button_text,
-          url: this.attachment.attributes.urls.edit_asset,
-          url_target: '_blank'
-        });
-      }
-      if (this.attachment.attributes.asset_type !== 'marvinjs'
-          && this.attachment.attributes.image_editable
-          && this.attachment.attributes.urls.start_edit_image) {
-        options.push({
-          text: this.i18n.t('assets.file_preview.edit_in_scinote'),
-          emit: 'open_scinote_editor'
-        });
-      }
-      if (this.canOpenLocally) {
-        const text = this.localAppName
-          ? this.i18n.t('attachments.open_locally_in', { application: this.localAppName })
-          : this.i18n.t('attachments.open_locally');
-
-        options.push({
-          text,
-          emit: 'open_locally',
-          data_e2e: 'e2e-BT-attachmentOptions-openLocally'
-        });
-      }
-      return options;
-    }
   }
 };
 </script>
